Change the ID to be a string. When using the file-based approach, the ID remains an Integer internally. However, in MongoDB we use the built-in ObjectId which can be represented as a hex string. This simplifies the process of giving new IDs to expenses (and would also theoretically better allow multiple servers to access the same database).

+++ src/main/kotlin/ch/cyril/budget/manager/backend/service/filebased/expense/ExpenseParser.kt
@@ -21,7 +21,7 @@ class ExpenseParser() : LineBasedFileParser<Expense>() {
 
     override fun fromLine(line: String): Expense {
         val split = line.split(",")
-        val id = Id(split[0].toInt())
+        val id = Id(split[0])
         val name = Name(split[1])
         val amount = Amount(split[2].toBigDecimal())
         val category = Category(split[3])
